LP Model for Optimal Generation Planning
The "RED_EVeC.lg4" is the LINGO 11 Set Code for generating the LP Model developed for determining the optimal power generation for a Renewable Energy Dominated electricity system when the demand for Electric VEhicles Charging is considered (hence the name "RED_EVeC"). The LP model determines the optimal power generation considering not only the cost of power generation but the social cost of emissions in power generation. Real-world operational aspects (such as ramp-up, base load, cold and hot starts, etc.) are also incorporated as constraints in the model.